requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
openai-agents
uvicorn
python-dotenv
pydantic